Increase font size of letter to 18px
I'm not sure if there's a Tailwind class we can use for this. I think that before the letter was small, which made people more likely to skip it. This is a small change but hopefully it stands a bit more. I also gave it some extra margin at the bottom to make it stand out.

I dropped "This year" to get that line all on one line.